ci-operator/step-registry/gather: Missing-kubeconfig gather guards
Sometimes we fail to even attempt to install a cluster, like [1]:

  $ curl -s https://storage.googleapis.com/origin-ci-test/pr-logs/pull/24838/pull-ci-openshift-origin-release-4.1-e2e-aws/370/build-log.txt | grep 'Executing\|[0-9] Container\|did not finish'
  2020/04/07 04:04:32 Executing "e2e-aws-ipi-conf"
  2020/04/07 04:04:37 Container cp-secret-wrapper in pod e2e-aws-ipi-conf completed successfully
  {"component":"entrypoint","file":"prow/entrypoint/run.go:164","func":"k8s.io/test-infra/prow/entrypoint.Options.ExecuteProcess","level":"error","msg":"Process did not finish before 4h0m0s timeout","time":"2020-04-07T07:38:29Z"}
  2020/04/07 07:38:29 Executing "e2e-aws-gather-must-gather"
  2020/04/07 07:38:34 Container cp-secret-wrapper in pod e2e-aws-gather-must-gather completed successfully
  2020/04/07 07:38:37 Container gather-must-gather in pod e2e-aws-gather-must-gather failed, exit code 1, reason Error
  2020/04/07 07:38:37 Executing "e2e-aws-gather-extra"
  2020/04/07 07:38:40 Container cp-secret-wrapper in pod e2e-aws-gather-extra completed successfully
  2020/04/07 07:38:43 Container gather-extra in pod e2e-aws-gather-extra failed, exit code 1, reason Error
  2020/04/07 07:38:43 Executing "e2e-aws-gather-audit-logs"
  2020/04/07 07:38:47 Container cp-secret-wrapper in pod e2e-aws-gather-audit-logs completed successfully
  2020/04/07 07:38:50 Container gather-audit-logs in pod e2e-aws-gather-audit-logs failed, exit code 1, reason Error
  2020/04/07 07:38:50 Executing "e2e-aws-ipi-deprovision-deprovision"
  2020/04/07 07:38:53 Container cp-secret-wrapper in pod e2e-aws-ipi-deprovision-deprovision completed successfully
  2020/04/07 07:39:15 Container ipi-deprovision-deprovision in pod e2e-aws-ipi-deprovision-deprovision failed, exit code 1, reason Error

When there is no kubeconfig, there's no sense in running all of this
logic and filling the container and build logs with distracting noise.
This commit makes gathers no-op successes when there is no kubeconfig,
making it easier to find jobs where we did have a running cluster but
failed to gather anyway.

There's still a window where we expect gathers to fail: when we
started off installing a cluster, but the install died early and the
cluster is missing or non-responsive.  But landing partial guards is
better than landing no guards at all ;).

[1]: https://prow.svc.ci.openshift.org/view/gcs/origin-ci-test/pr-logs/pull/24838/pull-ci-openshift-origin-release-4.1-e2e-aws/370

+++ ci-operator/step-registry/gather/audit-logs/gather-audit-logs-commands.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+
+if test ! -f "${KUBECONFIG}"
+then
+	echo "No kubeconfig, so no point in gathering audit logs."
+	exit 0
+fi
+
 dest_dir=$(mktemp --dir --tmpdir="$ARTIFACT_DIR")
 oc adm must-gather --dest-dir="$dest_dir" --image=quay.io/openshift/origin-must-gather -- /usr/bin/gather_audit_logs
 mv "$dest_dir"/*/audit_logs "$ARTIFACT_DIR/audit-logs"
